+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/conformanceMethod/domain/CertificationResultConformanceMethod.java
@@ -12,6 +12,7 @@ import org.apache.commons.lang3.StringUtils;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import gov.healthit.chpl.entity.listing.CertificationResultConformanceMethodEntity;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 
@@ -30,6 +31,7 @@ public class CertificationResultConformanceMethod implements Serializable {
     /**
      * Conformance Method to certification result mapping internal ID
      */
+    @Schema(description = "Conformance Method to certification result mapping internal ID")
     @XmlElement(required = true)
     private Long id;
 
@@ -38,6 +40,9 @@ public class CertificationResultConformanceMethod implements Serializable {
      * For the Test Procedure method, this also includes the version used during testing of the certification
      * criterion functionality.
      */
+    @Schema(description = "The method used to evaluate compliance with a certification criterion. It is applicable for 2015 Edition. "
+            + "For the Test Procedure method, this also includes the version used during testing of the certification "
+            + "criterion functionality.")
     @XmlElement(required = true)
     private ConformanceMethod conformanceMethod;
 
@@ -47,6 +52,10 @@ public class CertificationResultConformanceMethod implements Serializable {
      * This variable is a string variable that does not take any restrictions on
      * formatting or values and is applicable for 2015 Edition.
      */
+    @Schema(description = "The conformance method version used for a given certification criteria. "
+            + "It is valid for Test Procedure method only. "
+            + "This variable is a string variable that does not take any restrictions on "
+            + "formatting or values and is applicable for 2015 Edition.")
     @XmlElement(required = false)
     private String conformanceMethodVersion;