+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/InheritedCertificationStatus.java
@@ -14,10 +14,10 @@ import javax.xml.bind.annotation.XmlType;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.Singular;
 
 @XmlType(namespace = "http://chpl.healthit.gov/listings")
 @XmlAccessorType(XmlAccessType.FIELD)
@@ -45,6 +45,9 @@ public class InheritedCertificationStatus implements Serializable {
      * request. This variable is applicable for 2014 and 2015 Edition and a binary variable that takes either true or
      * false value.
      */
+    @Schema(description = "This variable indicates whether or not the certification issued was a result of an inherited certified status "
+            + "request. This variable is applicable for 2014 and 2015 Edition and a binary variable that takes either true or "
+            + "false value.")
     @XmlElement(name = "inherits")
     private Boolean inherits;
 
@@ -55,17 +58,19 @@ public class InheritedCertificationStatus implements Serializable {
     /**
      * The first-level parent listings that this listing inherits from
      */
+    @Schema(description = "The first-level parent listings that this listing inherits from")
     @XmlElementWrapper(name = "parents", nillable = true, required = false)
     @XmlElement(name = "parent")
-    @Singular
+    @Builder.Default
     private List<CertifiedProduct> parents = new ArrayList<CertifiedProduct>();
 
     /**
      * The first-level child listings that inherit from this listings
      */
+    @Schema(description = "The first-level child listings that inherit from this listings")
     @XmlElementWrapper(name = "children", nillable = true, required = false)
     @XmlElement(name = "child")
-    @Singular
+    @Builder.Default
     private List<CertifiedProduct> children = new ArrayList<CertifiedProduct>();
 
 }