+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/subscription/domain/ChplItemSubscription.java
@@ -1,24 +1,16 @@
 package gov.healthit.chpl.subscription.domain;
 
-import java.util.Date;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @Builder
-public class Subscription {
+public class ChplItemSubscription {
     private Long id;
-    private Subscriber subscriber;
     private SubscriptionSubject subject;
-    private Long subscribedObjectId;
     private SubscriptionConsolidationMethod consolidationMethod;
-    @JsonIgnore
-    private Date creationDate;
 }