+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/subscription/validation/SubscriptionRequestValidationService.java
@@ -21,7 +21,6 @@ public class SubscriptionRequestValidationService {
     private List<ValidationRule<SubscriptionRequestValidationContext>> getCreateValidations() {
         return new ArrayList<ValidationRule<SubscriptionRequestValidationContext>>(Arrays.asList(
                 new EmailAddressValidation(),
-                new SubscriberRoleValidation(),
                 new SubscriptionObjectTypeValidation(),
                 new SubscribedObjectValidation()));
     }